+++ src/main/java/com/cocoblue/securitytest/service/PostService.java
@@ -12,6 +12,6 @@ public interface PostService {
     public Boolean writePost(Post post);
     public void increaseViewNum(String post_id);
     public List<Post> getPostsByPage(String boardName, int page);
-    public int getPostsCount(String boardName);
+    public long getPostsCount(String boardName);
     public List<Post> getPostsByKeyword(String boardName, String keyword);
 }
